+++ sdk/resourcemanager/containerinstance/armcontainerinstance/models.go
@@ -23,6 +23,9 @@ type AzureFileVolume struct {
 
 	// The storage account access key used to access the Azure File share.
 	StorageAccountKey *string
+
+	// The reference to the storage account access key used to access the Azure File share.
+	StorageAccountKeyReference *string
 }
 
 // CachedImages - The cached image and OS type.
@@ -247,6 +250,9 @@ type ContainerGroupPropertiesProperties struct {
 	// The IP address type of the container group.
 	IPAddress *IPAddress
 
+	// The access control levels of the identities.
+	IdentityACLs *IdentityACLs
+
 	// The image registry credentials by which the container group is created from.
 	ImageRegistryCredentials []*ImageRegistryCredential
 
@@ -264,6 +270,9 @@ type ContainerGroupPropertiesProperties struct {
 
 	// The SKU for a container group.
 	SKU *ContainerGroupSKU
+
+	// The secret references that will be referenced within the container group.
+	SecretReferences []*SecretReference
 
 	// The subnet resource IDs for a container group.
 	SubnetIDs []*ContainerGroupSubnetID
@@ -460,6 +469,9 @@ type EnvironmentVariable struct {
 	// The value of the secure environment variable.
 	SecureValue *string
 
+	// The reference of the secure environment variable.
+	SecureValueReference *string
+
 	// The value of the environment variable.
 	Value *string
 }
@@ -544,6 +556,24 @@ type IPAddress struct {
 	Fqdn *string
 }
 
+// IdentityACLs - The access control levels of the identities.
+type IdentityACLs struct {
+	// The access control levels for each identity.
+	ACLs []*IdentityAccessControl
+
+	// The default access level.
+	DefaultAccess *IdentityAccessLevel
+}
+
+// IdentityAccessControl - The access control for an identity
+type IdentityAccessControl struct {
+	// The access level of the identity.
+	Access *IdentityAccessLevel
+
+	// An identity.
+	Identity *string
+}
+
 // ImageRegistryCredential - Image registry credential.
 type ImageRegistryCredential struct {
 	// REQUIRED; The Docker image registry server without a protocol such as "http" and "https".
@@ -557,6 +587,9 @@ type ImageRegistryCredential struct {
 
 	// The password for the private registry.
 	Password *string
+
+	// The reference for the private registry password.
+	PasswordReference *string
 
 	// The username for the private registry.
 	Username *string
@@ -733,6 +766,18 @@ type ResourceRequirements struct {
 	Limits *ResourceLimits
 }
 
+// SecretReference - A secret reference
+type SecretReference struct {
+	// REQUIRED; The ARM resource id of the managed identity that has access to the secret in the key vault
+	Identity *string
+
+	// REQUIRED; The identifier of the secret reference
+	Name *string
+
+	// REQUIRED; The URI to the secret in key vault
+	SecretReferenceURI *string
+}
+
 // SecurityContextCapabilitiesDefinition - The capabilities to add or drop from a container.
 type SecurityContextCapabilitiesDefinition struct {
 	// The capabilities to add to the container.
@@ -823,6 +868,9 @@ type Volume struct {
 
 	// The secret volume.
 	Secret map[string]*string
+
+	// The secret reference volume.
+	SecretReference map[string]*string
 }
 
 // VolumeMount - The properties of the volume mount.